requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
pyyaml>=6.0
pyarrow>=14.0
pytest>=7.4
